P3/OneOnOne/requirements.txt
Django
djangorestframework
rest_framework_simplejwt
djangorestframework_simplejwt